+++ src/model/nifmodel.cpp
@@ -2638,6 +2638,7 @@ void NifModel::updateLinks( int block, NifItem * parent )
 	if ( !parent )
 		return;
 
+	int n = 0;
 	for ( int l : parent->getLinkRows() ) {
 		NifItem * c = parent->child( l );
 		if ( !c )
@@ -2653,9 +2654,23 @@ void NifModel::updateLinks( int block, NifItem * parent )
 			if ( c->valueType() == NifValue::tUpLink ) {
 				if ( !parentLinks[block].contains( i ) )
 					parentLinks[block].append( i );
-			} else {
-				if ( !childLinks[block].contains( i ) )
-					childLinks[block].append( i );
+			} else if ( auto & cl = childLinks[block]; !cl.contains( i ) ) {
+				cl.append( i );
+
+				// Move the Data link to before Skin Instance and properties in NiParticleSystem (Skyrim SE and newer)
+				//   Row 25: Skin Instance
+				//   Row 26: Skin Instance
+				//   Row 29: Shader Property
+				//   Row 30: Alpha Property
+				//   Row 36: Data
+				if ( ( 1ull << l ) & 0x001066000000ull ) {
+					if ( l == 36 ) {
+						if ( n > 0 && bsVersion >= 100 && blockInherits( parent, "NiParticleSystem" ) )
+							cl.move( cl.size() - 1, cl.size() - ( n + 1 ) );
+					} else {
+						n++;
+					}
+				}
 			}
 		}
 	}